src/reviewservice/requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
grpcio-health-checking==1.59.3
protobuf==4.25.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
pydantic==2.5.0
google-cloud-secret-manager==2.17.0
google-cloud-logging==3.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-grpc==0.8.0
aiosqlite==0.19.0 